Added benchmarking tool (#52)
Removed a bunch of allocate and copy calls
Condensed most alloc/copy to areas in framework that can't be optimized yet
Added inlining for helper functions

+++ Kerberos.NET/Asn1/Experimental/AsnCharacterStringEncodings.cs
@@ -240,7 +240,7 @@ namespace System.Security.Cryptography.Asn1
             {
                 char c = chars[i];
 
-                if ((uint)c >= (uint)_isAllowed.Length || !_isAllowed[c])
+                if (c >= (uint)_isAllowed.Length || !_isAllowed[c])
                 {
                     EncoderFallback.CreateFallbackBuffer().Fallback(c, i);
